requirements.txt
numpy==1.26.4
python_igraph==0.11.4
scikit_learn==1.4.2
